requirements.txt
annotated-types==0.7.0
anyio==4.11.0
atlassian-python-api==4.0.7
attrs==25.4.0
Authlib==1.6.5
backports.tarfile==1.2.0
beartype==0.22.5
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
cyclopts==4.2.1
Deprecated==1.3.1
diskcache==5.6.3
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.2
email-validator==2.3.0
exceptiongroup==1.3.0
fastmcp==2.13.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.6.0
markdown-it-py==4.0.0
mcp==1.20.0
mdurl==0.1.2
more-itertools==10.8.0
oauthlib==3.3.1
openapi-pydantic==0.5.1
pathable==0.4.4
pathvalidate==3.3.1
platformdirs==4.5.0
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.11.0
python-dotenv==1.2.1
python-multipart==0.0.20
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.28.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.3
starlette==0.50.0
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
wrapt==2.0.0
zipp==3.23.0
